Begin refactoring file_reader to separate data readout, corrections and other actions to decouple AmorData class

+++ libeos/helpers_numba.py
@@ -11,7 +11,7 @@ def merge_frames(tof_e, tofCut, tau, total_offset):
         tof_e_out[ti] = ((tof_e[ti]-dt)%tau)+total_offset  # tof shifted to 1 frame
     return tof_e_out
 
-@nb.jit(nb.float64[:](nb.float64[:], nb.uint64[:], nb.int64[:]),
+@nb.jit(nb.int64[:](nb.float64[:], nb.uint64[:], nb.int64[:]),
         nopython=True, parallel=True, cache=True)
 def extract_walltime(tof_e, dataPacket_p, dataPacketTime_p):
     # assigning every event the wall time of the event packet (absolute time of pulse ?start?)
